PHOENIX-4835 LoggingPhoenixConnection should log metrics upon connection close
